requirements.txt
asgiref==3.8.1
attrs==24.3.0
blinker==1.8.2
certifi==2024.12.14
channels==4.1.0
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
distlib==0.3.8
Django==5.1.4
django-crispy-forms==2.3
django-model-utils==5.0.0
django-notifications-hq==1.8.3
django-rest-swagger==2.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
filelock==3.15.4
Flask==3.0.3
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
itypes==1.2.0
Jinja2==3.1.4
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mysqlclient==2.2.4
openapi==2.0.0
openapi-codec==1.3.2
packaging==24.2
pillow==10.4.0
platformdirs==4.2.2
psycopg2==2.9.9
PyJWT==2.9.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
simplejson==3.19.3
sqlparse==0.5.0
swapper==1.4.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.26.3
Werkzeug==3.0.4
